+++ pandas/tests/reshape/test_concat.py
@@ -2245,7 +2245,7 @@ def test_concat_empty_and_non_empty_frame_regression():
 def test_concat_empty_and_non_empty_series_regression():
     # GH 18187 regression test
     s1 = pd.Series([1])
-    s2 = pd.Series([])
+    s2 = pd.Series([], dtype='float')
     expected = s1
     result = pd.concat([s1, s2])
     tm.assert_series_equal(result, expected)
