+++ pandas/tests/series/test_replace.py
@@ -227,7 +227,7 @@ class TestSeriesReplace(TestData):
         # GH 15289
         s = pd.Series(list('abcd'))
         tm.assert_series_equal(s, s.replace(dict()))
-        tm.assert_series_equal(s, s.replace(pd.Series([])))
+        tm.assert_series_equal(s, s.replace(pd.Series([], dtype='float')))
 
     def test_replace_string_with_number(self):
         # GH 15743
