+++ pandas/tests/test_algos.py
@@ -585,7 +585,8 @@ class TestIsin(object):
         result = algos.isin(Sd, St)
         tm.assert_numpy_array_equal(expected, result)
 
-    @pytest.mark.parametrize("empty", [[], Series(), np.array([])])
+    @pytest.mark.parametrize("empty", [[], Series(dtype='float'),
+                             np.array([])])
     def test_empty(self, empty):
         # see gh-16991
         vals = Index(["a", "b"])
