+++ pandas/tests/util/test_hashing.py
@@ -152,7 +152,7 @@ class TestHashing(object):
                     Series(['a', np.nan, 'c']),
                     Series(['a', None, 'c']),
                     Series([True, False, True]),
-                    Series(),
+                    Series(dtype='float'),
                     Index([1, 2, 3]),
                     Index([True, False, True]),
                     DataFrame({'x': ['a', 'b', 'c'], 'y': [1, 2, 3]}),
